move syncAndStopNode to different file
